adding testing target and test cases for GPXDocumentMerge
refactored GPXDocumentMerge to make the utility functions type methods.
this enables easier testing of the functions, since they do not rely
on state maintained by an instance of the class.

however, these functions still *do* mutate the types passed as
parameters, because of how the underlying NSXMLNode methods work.
it would be more functional to *always* return new instances
of the passed in NSXMLNode objects.